1. registry files corresponding to the recent registry.h changes 2. badly famous wxRegConfig::DeleteAll() bug corrected
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@1456 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/regconf.cpp
@@ -528,16 +528,15 @@ bool wxRegConfig::DeleteAll()
 {
   m_keyLocal.Close();
   m_keyGlobal.Close();
-#if 1
-  wxFAIL_MSG("wxRegConfig::DeleteAll will wipe out your entire registry, so please do not use until it's fixed!");
-  return FALSE;
-#else
+
   bool bOk = m_keyLocalRoot.DeleteSelf();
-  if ( bOk )
+
+  // make sure that we opened m_keyGlobalRoot and so it has a reasonable name:
+  // otherwise we will delete HKEY_CLASSES_ROOT recursively
+  if ( bOk && m_keyGlobalRoot.IsOpened() )
     bOk = m_keyGlobalRoot.DeleteSelf();
 
   return bOk;
-#endif
 }
 
 #endif
